backend/requirements.txt
fastapi>=0.111.0
ultralytics>=8.2.0
opencv-python-headless>=4.10.0.84
grad-cam>=1.5.2
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.4.0
loguru>=0.7.2
slowapi>=0.1.9
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
supabase>=2.4.2
httpx>=0.27.0
torch>=2.3.0
torchvision>=0.18.0
